+++ javascript/node/selenium-webdriver/lib/http.js
@@ -260,10 +260,8 @@ const W3C_COMMAND_MAP = new Map([
   [cmd.Name.MAXIMIZE_WINDOW, post('/session/:sessionId/window/maximize')],
   [cmd.Name.MINIMIZE_WINDOW, post('/session/:sessionId/window/minimize')],
   [cmd.Name.FULLSCREEN_WINDOW, post('/session/:sessionId/window/fullscreen')],
-  [cmd.Name.GET_WINDOW_POSITION, get('/session/:sessionId/window/position')],
-  [cmd.Name.SET_WINDOW_POSITION, post('/session/:sessionId/window/position')],
-  [cmd.Name.GET_WINDOW_SIZE, get('/session/:sessionId/window/size')],
-  [cmd.Name.SET_WINDOW_SIZE, post('/session/:sessionId/window/size')],
+  [cmd.Name.GET_WINDOW_RECT, get('/session/:sessionId/window/rect')],
+  [cmd.Name.SET_WINDOW_RECT, post('/session/:sessionId/window/rect')],
   [cmd.Name.GET_CURRENT_WINDOW_HANDLE, get('/session/:sessionId/window')],
   [cmd.Name.GET_WINDOW_HANDLES, get('/session/:sessionId/window/handles')],
 ]);